+++ TheCutestWayToDie-core/src/com/thedistrictheat/gameobjects/EnemyJumping.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.math.Intersector;
 public class EnemyJumping extends Enemy{
     private static final int ENEMY_JUMPING_SPEED = -22;
 
-	public EnemyJumping(float x, float y, int width, int height) {
+	public EnemyJumping(int x, int y, int width, int height) {
 		super(x, y, width, height, ENEMY_JUMPING_SPEED, EnemyType.JUMPING);
 	}
