+++ TheCutestWayToDie-core/src/com/thedistrictheat/gameobjects/EnemyWalking.java
@@ -3,7 +3,7 @@ package com.thedistrictheat.gameobjects;
 public class EnemyWalking extends Enemy{
     private static final int ENEMY_WALKING_SPEED = -30;
 
-	public EnemyWalking(float x, float y, int width, int height) {
+	public EnemyWalking(int x, int y, int width, int height) {
 		super(x, y, width, height, ENEMY_WALKING_SPEED, EnemyType.WALKING);
 	}
